增加了世界各数据集合合约 增加角色一些属性合约 增加了一套出生序列事件合约 增加测试用的预设数据文件和加载脚本 增加一个角色出生序列测试用例

+++ packages/taiyi-contracts/contracts/world/events/WorldEventProcessor10001.sol
@@ -17,18 +17,18 @@ check order:
 
 contract WorldEventProcessor10001 is DefaultWorldEventProcessor {
 
-    uint[] public maleActors;
+    uint256[] public maleActors;
 
     constructor(address _worldRouteAddress) DefaultWorldEventProcessor(_worldRouteAddress, 60002) {
     }
 
-    function eventInfo(uint /*_actor*/) external virtual view override returns (string memory) {
+    function eventInfo(uint256 /*_actor*/) external virtual view override returns (string memory) {
         //你出生了，是个男孩。
         return "\xE4\xBD\xA0\xE5\x87\xBA\xE7\x94\x9F\xE4\xBA\x86\xEF\xBC\x8C\xE6\x98\xAF\xE4\xB8\xAA\xE7\x94\xB7\xE5\xAD\xA9\xE3\x80\x82";
     }
     //乾
-    function trigrams(uint /*_actor*/) virtual external override view returns (uint[] memory) {
-        uint[] memory _t = new uint[](6);
+    function trigrams(uint256 /*_actor*/) virtual external override view returns (uint[] memory) {
+        uint[] memory _t = new uint256[](6);
         _t[0] = 1;
         _t[1] = 1;
         _t[2] = 1;
@@ -37,26 +37,26 @@ contract WorldEventProcessor10001 is DefaultWorldEventProcessor {
         _t[5] = 1;
         return _t;
     }
-    function checkOccurrence(uint _actor, uint /*_age*/) external view override returns (bool) {
+    function checkOccurrence(uint256 _actor, uint256 /*_age*/) external view override returns (bool) {
         bool defaultRt = true;
 
         IActorTalents talents = IActorTalents(worldRoute.modules(WorldConstants.WORLD_MODULE_TALENTS));
 
         //"exclude": "TLT?[1004,1024,1025]"
-        uint[] memory tlts = talents.actorTalents(_actor);
-        for(uint i=0; i<tlts.length; i++) {
+        uint256[] memory tlts = talents.actorTalents(_actor);
+        for(uint256 i=0; i<tlts.length; i++) {
             if(tlts[i] == 1004 || tlts[i] == 1024 || tlts[i] == 1025)
                 return false;
         }
 
         return defaultRt;
     }
-    function process(uint256 _operator, uint _actor, uint /*_age*/) external override 
+    function process(uint256 _operator, uint256 _actor, uint256 /*_age*/) external override 
         onlyYeMing(_operator)
     {
         maleActors.push(_actor);
     }
-    function maleNum() external view returns (uint) {
+    function maleNum() external view returns (uint256) {
         return maleActors.length;
     }
 }
